% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.rrt_repo}
\alias{as.rrt_repo}
\alias{is.rrt_repo}
\title{Functions to check if an object is a rrt_repo, or coerce it if possible.}
\usage{
is.rrt_repo(x)

as.rrt_repo(x)
}
\arguments{
\item{x}{A list to coerce into an rrt_repo object}
}
\description{
Functions to check if an object is a rrt_repo, or coerce it if possible.
}

