% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpoint_remove.R
\name{checkpointArchives}
\alias{checkpointArchives}
\title{List checkpoint archives on disk.}
\usage{
checkpointArchives(checkpointLocation = "~/")
}
\arguments{
\item{checkpointLocation}{File path where the checkpoint library is stored.  Default is \code{"~/"}, i.e. the user's home directory. A use case for changing this is to create a checkpoint library on a portable drive (e.g. USB drive).}
}
\description{
List checkpoint archives on disk.
}
\examples{
checkpointArchives()
\dontrun{
checkpointRemove("2016-10-01")
}

}
\seealso{
checkpointRemove
}
