% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createRepoFolders}
\alias{createRepoFolders}
\title{Create all required repo folders.}
\usage{
createRepoFolders(repo)
}
\arguments{
\item{repo}{Repo}
}
\description{
Create all required repo folders.
}
\seealso{
Other Repo path: \code{\link{rrtPath}}
}

