% Generated by roxygen2 (4.0.1): do not edit by hand
\name{download_one_github_pkg}
\alias{download_one_github_pkg}
\title{Install packages from locations other than CRAN, its mirrors, or MRAN}
\usage{
download_one_github_pkg(pkg, username, lib, ...)
}
\description{
Install packages from locations other than CRAN, its mirrors, or MRAN
}
\examples{
\dontrun{
download_one_github_pkg(lib=tempdir(), pkg="cowsay", username="sckott")
dir(file.path(tempdir, "src/contrib"))
}
}
\keyword{internal}

