% Generated by roxygen2 (4.0.1): do not edit by hand
\name{download_pkgs_mran}
\alias{download_pkgs_mran}
\title{Download R packages from the MRAN server}
\usage{
download_pkgs_mran(repo = NULL, lib = NULL, date = NULL,
  snapshotid = NULL, pkgs = NULL, outdir = NULL, verbose = FALSE,
  quiet = FALSE)
}
\arguments{
\item{repo}{Repository path}

\item{lib}{(character) Library location, a directory}

\item{date}{Date as "year-month-day" (YY-MM-DD)}

\item{snapshotid}{Optional. You can give the exact snapshot ID insetad of a date.}

\item{outdir}{Output directory}

\item{pkgs}{Packages to install with version numbers, e.g. plyr_1.8.1}

\item{quiet}{Passed to \code{\link[utils]{install.packages}}}

\item{verbose}{(logical) Whether to print messages or not (Default: FALSE).}
}
\description{
This function uses rsync on *unix machines, which is faster than the method (wget) \code{install.packages} uses by default. On Windows we use your default method of downloading files. This function does not install packages, but only downloads them to your machine.
}
\examples{
\dontrun{
# By default installs most recent version
download_pkgs_mran(date='2014-08-04', pkgs=c("plyr","ggplot2"), outdir="~/mran_snaps/")

download_pkgs_mran(date='2014-06-19', pkgs=c("plyr_1.8.1","ggplot2_1.0.0"), outdir="~/mran_snaps/stuff/")
download_pkgs_mran(date='2014-06-19', pkgs="rgbif_0.6.2", outdir="~/mran_snaps")
}
}

