% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_github}
\alias{get_github}
\title{Install packages from locations other than CRAN, its mirrors, or MRAN}
\usage{
get_github(pkg, username, lib, ...)
}
\description{
Install packages from locations other than CRAN, its mirrors, or MRAN
}
\examples{
\dontrun{
get_github(lib=lib, pkg="cowsay", username="sckott")
}
}
\keyword{internal}

