% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_github_pkgs}
\alias{get_github_pkgs}
\title{Download github packages.}
\usage{
get_github_pkgs(repo, pkgs, lib)
}
\arguments{
\item{repo}{Repository path}

\item{pkgs}{Vector of package names}

\item{lib}{Library path}
}
\description{
Download github packages.
}
\keyword{internal}

