% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getSnapshotUrl}
\alias{getSnapshotUrl}
\title{Retrieves snapshot id for a given snapshot date.}
\usage{
getSnapshotUrl(snapshotdate, autosnap = c("first", "last", "all"),
  url = mranUrl(), returnUrl = TRUE)
}
\arguments{
\item{url}{MRAN snapshot URL}

\item{returnUrl}{If TRUE, returns URL with snapshotdate appended, otherwise returns snapshotdate as a string}

\item{snapshotdate}{(date) Required. Date of snapshot to use. E.g. "2014-06-20". If left blank, you
will be supplied with options.}

\item{autosnap}{One of first, last or all. Determines how to resolve the snapshot id for a given date, if more than one snapshot exists on MRAN for that date.}
}
\description{
Retrieves snapshot id for a given snapshot date.
}
\examples{
getSnapshotUrl()
getSnapshotUrl(autosnap = "first")
getSnapshotUrl(autosnap = "last")
getSnapshotUrl(autosnap = "all")
getSnapshotUrl("2014-08-18")
getSnapshotUrl("2014-09-08")
getSnapshotUrl("2014-09-08", autosnap = "last")
}

