% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mranUrl.R
\name{getValidSnapshots}
\alias{getValidSnapshots}
\title{Read list of available snapshot dates from MRAN url.}
\usage{
getValidSnapshots(url = mranUrl())
}
\description{
Read list of available snapshot dates from MRAN url.
}

