% Generated by roxygen2 (4.0.1): do not edit by hand
\name{git_add}
\alias{git_add}
\title{Add and commit files in your RRT repository}
\usage{
git_add(path = getwd(), verbose = TRUE)
}
\arguments{
\item{path}{Path to the RRT repository}

\item{verbose}{Print messages or not, Default: TRUE}
}
\description{
Add and commit files in your RRT repository
}
\examples{
\dontrun{
git_add(path="~/bbb")
}
}

