% Generated by roxygen2 (4.0.1): do not edit by hand
\name{git_setup}
\alias{git_setup}
\title{Add and commit files in your RRT repository}
\usage{
git_setup(path = getwd(), verbose = TRUE, ...)
}
\arguments{
\item{path}{Path to the RRT repository}

\item{verbose}{Print messages or not, Default: TRUE}

\item{...}{Further args, ignored for now.}
}
\description{
Add and commit files in your RRT repository
}
\examples{
\dontrun{
git_add(path="~/bbb")
}
}

