% Generated by roxygen2 (4.0.1): do not edit by hand
\name{install_mran}
\alias{install_mran}
\title{Install packages from MRAN}
\usage{
install_mran(pkg, repo, date = NULL, libPath = rrtPath(repo, "lib"),
  srcPath = rrtPath(repo, "src"), snapshot = getSnapshotId(date),
  quiet = FALSE, ..., dependencies = TRUE)
}
\arguments{
\item{pkg}{Package name}

\item{date}{MRAN snapshot date}

\item{libPath}{Library to install packages in}

\item{srcPath}{Destination directory to install packages to}

\item{...}{Further args passed on to \code{\link[utils]{install.packages}}}

\item{dependencies}{Whether to install dependencies or not. Default: TRUE. See}

\item{quiet}{Passed to \code{\link[utils]{install.packages}}
\link{install.packages}}
}
\value{
Installs a package, or throws warnings/errors on failures
}
\description{
Steps (for each pkg):
 download package
 determine dependencies
 download dependencies (but don't install them)
 install target pkg and downloaded dependencies
}
\examples{
\dontrun{
  repo_path <- tempdir()
  createRepoFolders(repo_path)
  
  install_mran(repo=repo_path, pkg="plyr", date="2014-08-01")
  
  dir(instPath)
  unlink(instPath, recursive = TRUE)
}



}
\seealso{
Other mran: \code{\link{mranPkgMetadata}};
  \code{\link{mranPkgVersions}};
  \code{\link{mranSnapshots}}; \code{\link{mran_diffs}}
}

