% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{install_other}
\alias{install_other}
\title{Install packages that were downloaded from non-CRAN like places}
\usage{
install_other(pkg, lib)
}
\arguments{
\item{pkg}{Package name}

\item{lib}{Path to packages library}
}
\description{
Install packages that were downloaded from non-CRAN like places
}
\examples{
\dontrun{
install_other(pkg="cowsay", lib=lib)
}
}
\keyword{internal}

