% Generated by roxygen2 (4.0.1): do not edit by hand
\name{installPkgs}
\alias{installPkgs}
\title{Function to install pkgs}
\usage{
installPkgs(x, lib, recursive = FALSE)
}
\arguments{
\item{x}{}

\item{lib}{}

\item{recursive}{}
}
\description{
Function to install pkgs
}
\examples{
\dontrun{
installPkgs()
}
}

