% Generated by roxygen2 (4.0.1): do not edit by hand
\name{installRepoPackages}
\alias{installRepoPackages}
\title{Installs downloaded repo packages from source.}
\usage{
installRepoPackages(repo, libPath = rrtPath(repo, "lib"),
  srcPath = rrtPath(repo, "src"), pkgs, verbose, quiet = FALSE)
}
\description{
Installs downloaded repo packages from source.
}

