% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is_bioc_pkg}
\alias{is_bioc_pkg}
\title{Determine if a package is on Bioconductor or not.}
\usage{
is_bioc_pkg(pkgs)
}
\arguments{
\item{pkgs}{Vector of package names}
}
\value{
A logical vector of same length as input vector
}
\description{
Determine if a package is on Bioconductor or not.
}
\keyword{internal}

