% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{is_rrt}
\alias{is_rrt}
\title{Find out if a repository is an RRT repository.}
\usage{
is_rrt(repo = getwd(), verbose = TRUE)
}
\arguments{
\item{repo}{(character) A path to create a RRT repository; defaults to current working directory}

\item{verbose}{Print messages or not, Default: TRUE}
}
\value{
A logical, TRUE or FALSE
}
\description{
Find out if a repository is an RRT repository.
}
\examples{
\dontrun{
is_rrt(repo="~/testrepo")
}
}

