% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mran_utils.R
\name{list_mran_snapshots}
\alias{list_mran_snapshots}
\title{Lists the snapshot dates found on MRAN}
\usage{
list_mran_snapshots(
  mran_url = getOption("checkpoint.mranUrl", "https://mran.microsoft.com")
)
}
\arguments{
\item{mran_url}{The base MRAN URL. The default is taken from the system option \code{checkpoint.mranUrl}, or if this is unset, \verb{https://mran.microsoft.com}.}
}
\value{
A character vector of snapshot dates.
}
\description{
Lists the snapshot dates found on MRAN
}
