% Generated by roxygen2 (4.0.1): do not edit by hand
\name{marmoset_diffs}
\alias{marmoset_diffs}
\title{Get available diffs from Marmoset}
\usage{
marmoset_diffs(diff = NULL)
}
\arguments{
\item{diff}{Optional. (character) A diff date-time stamp of a Marmoset diff.}
}
\description{
Get available diffs from Marmoset
}
\examples{
\dontrun{
marmoset_diffs()

# An individual diff
marmoset_diffs(diff="2014-06-19_0136")
}
}

