% Generated by roxygen2 (4.0.1): do not edit by hand
\name{marmoset_pkg_avail}
\alias{marmoset_pkg_avail}
\title{Get available package versions from Marmoset}
\usage{
marmoset_pkg_avail(snapshot = NULL, package)
}
\arguments{
\item{snapshot}{A Marmoset snapshot. Defaults to most recent snapshot}

\item{package}{Required. A package name}
}
\description{
Get available package versions from Marmoset
}
\examples{
\dontrun{
marmoset_pkg_avail(snapshot="2014-06-19_0136", package="plyr")
}
}

