% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mran_diffs}
\alias{mran_diffs}
\title{Get available diffs from MRAN}
\usage{
mran_diffs(diff = NULL, which = "src", os = "macosx")
}
\arguments{
\item{diff}{Optional. (character) A diff date-time stamp of a MRAN diff.}

\item{which}{(character) One of src (for source packages) or bin (for binary packages).}

\item{os}{(character) Operating system. One of macosx, windows, or linux.}
}
\description{
Get available diffs from MRAN
}
\examples{
\dontrun{
mran_diffs()
mran_diffs(which='bin')
mran_diffs(which='bin', os='windows')
mran_diffs(which='bin', os='linux')

# An individual diff
mran_diffs(diff="2014-08-01_0500")

diffs <- mran_diffs()
mran_diffs(diffs[length(diffs)-1])
}
}
\seealso{
Other mran: \code{\link{install_mran}};
  \code{\link{mranPkgMetadata}};
  \code{\link{mranPkgVersions}};
  \code{\link{mranSnapshots}}
}

