% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mran_pkg_avail}
\alias{mran_pkg_avail}
\title{Get available package versions from MRAN}
\usage{
mran_pkg_avail(snapshot = NULL, package)
}
\arguments{
\item{snapshot}{A MRAN snapshot. Defaults to most recent snapshot}

\item{package}{Required. A package name}
}
\description{
Get available package versions from MRAN
}
\examples{
\dontrun{
mran_pkg_avail(snapshot="2014-06-19_0136", package="plyr")

# Example of differences in available versions between snapshots for the package MPSEM
snaps <- mran_snaps()
mran_pkg_avail(snaps[length(snaps)-1], package="MPSEM")
mran_pkg_avail(snaps[length(snaps)-2], package="MPSEM")
}
}

