% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mran_pkg_versions}
\alias{mran_pkg_versions}
\title{Get available package versions from MRAN}
\usage{
mran_pkg_versions(package, snapshot = NULL, which = "src", os = "macosx")
}
\arguments{
\item{package}{(character) Required. A package name}

\item{snapshot}{(date) An MRAN snapshot ('YYYY-MM-DD_TTTT') or a date ('YYYY-MM-DD'). Defaults to most
recent snapshot.}

\item{which}{(character) One of src or bin}

\item{os}{(character) Operating system. One of 'macosx', 'linux', or 'windows'}
}
\description{
Get available package versions from MRAN
}
\examples{
\dontrun{
mran_pkg_versions(snapshot="2014-07-14", package="plyr")
mran_pkg_versions(snapshot="2014-08-04", package="plyr", which="bin", os="windows")
}
}
\seealso{
Other mran: \code{\link{install_mran}};
  \code{\link{mran_diffs}};
  \code{\link{mran_pkg_metadata}}; \code{\link{mran_set}};
  \code{\link{mran_snaps}}
}

