% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mran_set}
\alias{mran_set}
\title{Set the MRAN snapshot to get packages from.}
\usage{
mran_set(snapshot = NULL, repo = getwd())
}
\arguments{
\item{snapshot}{A MRAN snapshot. Defaults to most recent snapshot}

\item{repo}{Repository path.}
}
\description{
Set the MRAN snapshot to get packages from.
}
\examples{
\dontrun{
mran_set()
mran_set(snapshot="2014-07-01_2000")
}
}
\seealso{
Other mran: \code{\link{install_mran}};
  \code{\link{mran_diffs}};
  \code{\link{mran_pkg_metadata}};
  \code{\link{mran_pkg_versions}}; \code{\link{mran_snaps}}
}

