% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mran_snaps}
\alias{mran_snaps}
\title{Get available snapshots from MRAN}
\usage{
mran_snaps(date = NULL)
}
\arguments{
\item{date}{(character) A date, in the format YYY-MM-DD}
}
\description{
Get available snapshots from MRAN
}
\examples{
\dontrun{
# List all available snapshots
mran_snaps()
# Get code for a single snapshot
mran_snaps(date='2014-08-04')
}
}
\seealso{
Other mran: \code{\link{install_mran}};
  \code{\link{mran_diffs}};
  \code{\link{mran_pkg_metadata}};
  \code{\link{mran_pkg_versions}}; \code{\link{mran_set}}
}

