% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mranPkgMetadata}
\alias{mranPkgMetadata}
\title{Get available package level metadata from MRAN.}
\usage{
mranPkgMetadata(package, snapshot = NULL)
}
\arguments{
\item{package}{Required. A package name}

\item{snapshot}{An MRAN snapshot ('YYYY-MM-DD_TTTT') or a date ('YYYY-MM-DD'). Defaults to most recent snapshot.}
}
\description{
Get available package level metadata from MRAN.
}
\examples{
\dontrun{
mranPkgMetadata(package="plyr", snapshot="2014-08-04")
}
}
\seealso{
Other mran: \code{\link{install_mran}};
  \code{\link{mranSnapshots}}; \code{\link{mran_diffs}};
  \code{\link{pkgVersionAtSnapshot}}
}

