% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mranPkgVersions}
\alias{mranPkgVersions}
\title{Get available package versions from MRAN}
\usage{
mranPkgVersions(package, snapshot = NULL, type = c("src", "mac.binary",
  "win.binary"))
}
\arguments{
\item{package}{(character) Required. A package name}

\item{snapshot}{(date) An MRAN snapshot ('YYYY-MM-DD_TTTT') or a date ('YYYY-MM-DD'). Defaults to most recent snapshot.}

\item{type}{(character) "src", "mac.binary" or "win.binary"}
}
\description{
Get available package versions from MRAN
}
\examples{
\dontrun{
mranPkgVersions(snapshot="2014-07-14", package="plyr")
mranPkgVersions(snapshot="2014-08-04", package="plyr", which="bin", os="windows")
}
}
\seealso{
Other mran: \code{\link{install_mran}};
  \code{\link{mranPkgMetadata}};
  \code{\link{mranSnapshots}}; \code{\link{mran_diffs}}
}

