% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mranSnapshots}
\alias{mranSnapshots}
\title{Get available snapshots from MRAN}
\usage{
mranSnapshots(snapshotdate = NULL, verbose = TRUE)
}
\arguments{
\item{snapshotdate}{(date) Required. Date of snapshot to use. E.g. "2014-06-20". If left blank, you
will be supplied with options.}

\item{verbose}{(logical) Whether to print messages or not (Default: TRUE).}
}
\description{
Get available snapshots from MRAN
}
\examples{
\dontrun{
# List all available snapshots
mranSnapshots()
# Get code for a single snapshot
mranSnapshots(date='2014-08-04')
}
}
\seealso{
Other mran: \code{\link{install_mran}};
  \code{\link{mranPkgMetadata}}; \code{\link{mran_diffs}};
  \code{\link{pkgVersionAtSnapshot}}
}

