% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mranUrl.R
\name{mranUrl}
\alias{mranUrl}
\title{Returns MRAN URL by querying options and defaults.}
\usage{
mranUrl()
}
\value{
Character string with URL
}
\description{
The default MRAN URL is \code{http(s)://mran.microsoft.com/}, but you can override this by setting the \code{checkpoint.mranUrl} option.
}
\seealso{
Other checkpoint functions: \code{\link{checkpointArchives}},
  \code{\link{checkpointRemove}}, \code{\link{checkpoint}},
  \code{\link{getValidSnapshots}},
  \code{\link{setSnapshot}}
}
