% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pkg_deps}
\alias{pkg_deps}
\title{Determine packages dependencies.}
\usage{
pkg_deps(pkg = ".", dependencies = NA)
}
\arguments{
\item{pkg}{package description, can be path or package name. See \code{\link[devtools]{as.package}} for more information}

\item{dependencies}{logical indicating to also install uninstalled packages which this \code{pkg} depends on/links to/suggests. See argument \code{dependencies} of \code{\link[utils]{install.packages}}.}
}
\description{
Determine packages dependencies.
}
\note{
This function copied from devtools:::pkg_deps
}
\keyword{Internal}

