% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pkgAvail}
\alias{pkgAvail}
\title{Reads available packages from CRAN repository.}
\usage{
pkgAvail(repos = getOption("repos"), type = getOption("pkgType"), ...)
}
\arguments{
\item{repos}{URL(s) of the 'contrib' sections of the repositories. Passed to \code{\link{available.packages}}}

\item{type}{Passed to \code{\link{available.packages}}}

\item{...}{Further args, not used}
}
\description{
This is a thin wrapper around \code{\link{available.packages}}.  If the argument \code{path} is
supplied, then the function attempts to read from a local repository, otherwise attempts to read
from a CRAN mirror at the \code{repos} url.
}

