% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pkgs_marmoset}
\alias{pkgs_marmoset}
\title{Install from Marmoset server}
\usage{
pkgs_marmoset(date = NULL, user = "sckott", outdir = NULL, pkgs = NULL)
}
\arguments{
\item{date}{Date as "year-month-day" (YY-MM-DD)}

\item{user}{User name}

\item{outdir}{Output directory}

\item{pkgs}{Packages to install with version numbers, e.g. plyr_1.8.1}
}
\description{
Install from Marmoset server
}
\examples{
\dontrun{
pkgs_marmoset(date='2014-06-17', user='sckott', outdir="~/marmoset_snaps/stuff/",
   pkgs=c("plyr_1.8.1","ggplot2_1.0.0"))
pkgs_marmoset(date='2014-06-17', user='sckott', outdir="~/marmoset_snaps/stuff/",
   pkgs="rgbif_0.6.2")
}
}

