% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pkgs_mran}
\alias{pkgs_mran}
\title{Install from MRAN server}
\usage{
pkgs_mran(date = NULL, pkgs = NULL, outdir = NULL)
}
\arguments{
\item{date}{Date as "year-month-day" (YY-MM-DD)}

\item{outdir}{Output directory}

\item{pkgs}{Packages to install with version numbers, e.g. plyr_1.8.1}
}
\description{
Install from MRAN server
}
\examples{
\dontrun{
# By default installs most recent version
pkgs_mran(date='2014-06-19', pkgs=c("plyr","ggplot2"), outdir="~/mran_snaps/stuff/")

pkgs_mran(date='2014-06-19', pkgs=c("plyr_1.8.1","ggplot2_1.0.0"), outdir="~/mran_snaps/stuff/")
pkgs_mran(date='2014-06-19', pkgs="rgbif_0.6.2", outdir="~/mran_snaps/stuff/")
}
}

