% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pkgurls}
\alias{pkgurls}
\title{Get package URIs given package name and version - for archived packages.}
\usage{
pkgurls(pkgs)
}
\arguments{
\item{pkgs}{A data.frame with two columns, or a list or vector with structure \code{pkgname_ver}}
}
\value{
A vector of urls of length equal to rows in the input data.frame
}
\description{
Get package URIs given package name and version - for archived packages.
}
\examples{
(dat <- data.frame(pkg=c('plyr','reshape2','shiny'), ver=c('1.6.0','1.2.1','0.8.0')))
pkgurls(pkgs=dat)
pkgurls(pkgs=c('plyr_1.6.0','reshape2_1.2.1','shiny_0.8.0'))
}

