% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pkgVersionAtSnapshot}
\alias{pkgVersionAtSnapshot}
\title{Get available package versions from MRAN}
\usage{
pkgVersionAtSnapshot(pkgs, snapshotdate,
  snapshotid = snapshotFromDate(snapshotdate), type = c("src", "mac.binary",
  "win.binary"), Rversion = "R3.0")
}
\arguments{
\item{pkgs}{(character) Required. A package name}

\item{type}{(character) "src", "mac.binary" or "win.binary"}

\item{Rversion}{not yet used}

\item{snapshotdate}{(date) Required. Date of snapshot to use. E.g. "2014-06-20". If left blank, you
will be supplied with options.}

\item{snapshotid}{(character) The identifying snapshot id, consisting of a date and semi-timestamp, e.g. "2014-08-01_0500"}
}
\description{
Get available package versions from MRAN
}
\examples{
pkgVersionAtSnapshot(snapshotdate="2014-07-14", pkgs="plyr")
pkgVersionAtSnapshot(snapshotdate="2014-08-01", pkgs="plyr", type="win.binary")
}
\seealso{
Other mran: \code{\link{install_mran}};
  \code{\link{mranPkgMetadata}};
  \code{\link{mranSnapshots}}; \code{\link{mran_diffs}}
}

