% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repoDigest}
\alias{repoDigest}
\title{Creates unique repo id from a digest of the file path.}
\usage{
repoDigest(repo)
}
\arguments{
\item{repo}{A repository path. This is the path to the root of your RRT repository. Defaults to current working directory current working directory via /code{/link{getwd}}.}
}
\description{
Creates unique repo id from a digest of the file path.
}
\keyword{Internal}

