% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repoInstalledPackages}
\alias{repoInstalledPackages}
\title{Return matrix of DESCRIPTION information of all packages installed in repo.}
\usage{
repoInstalledPackages(repo, libPath = rrtPath(repo, "lib"))
}
\arguments{
\item{libPath}{(character) Location of package library in repo}

\item{repo}{A repository path. This is the path to the root of your RRT repository. Defaults to current working directory current working directory via \code{\link{getwd}}.}
}
\description{
Return matrix of DESCRIPTION information of all packages installed in repo.
}

