% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repoScanPackages}
\alias{repoScanPackages}
\title{Detect repository packages.}
\usage{
repoScanPackages(repo = getwd(), verbose = TRUE)
}
\arguments{
\item{repo}{A repository path. This is the path to the root of your RRT repository. Defaults to current working directory current working directory via /code{/link{getwd}}.}

\item{verbose}{(logical) Whether to print messages or not (Default: TRUE).}
}
\value{
Vector of package names on which R code in the repository depends.
}
\description{
Detect all packages used in repository. Parses all .R, .Rmd, .Rnw, .Rpres, and other files in the repository directory to determine what packages the repo depends directly.
}
\details{
Only direct dependencies are detected (i.e. no recursion).

Packages are determined by parsing repository source code and looking for calls to \code{library}, \code{require}, \code{::}, and \code{:::}.
}
\note{
Modified from dependencies.R in the packrat github repo
}

