% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rrt}
\alias{rrt}
\title{Initiate or refresh an RRT repository.}
\usage{
rrt(repo = getwd(), mran = TRUE, snapdate = NULL, autosnap = FALSE,
  verbose = TRUE, rprofile = NULL, interactive = FALSE,
  suggests = FALSE)
}
\arguments{
\item{repo}{(character) A path to create a RRT repository; defaults to current working directory.}

\item{mran}{(logical) If TRUE (default), packages are installed from the MRAN server. See
\url{http://mran.revolutionanalytics.com} for more information.}

\item{snapdate}{Date of snapshot to use. E.g. "2014-06-20"}

\item{autosnap}{(logical) Get most recent snapshot. Default: FALSE}

\item{verbose}{(logical) Whether to print messages or not (Default: TRUE).}

\item{rprofile}{(list) pass in a list of options to include in the .Rprofile file for the repo.}

\item{interactive}{(logical) If TRUE, function asks you for input for each item,
otherwise, defaults are used. Default: FALSE.}

\item{suggests}{(logical) Download and install packages in the Suggests line for packages
used in your RRT repository, or not. Default: FALSE.}
}
\value{
Files written to the user's machine, with informative messages on progress
}
\description{
This function decides what to do with your repository based on commands you give, and on
skimming through your repo. You can run this function to start a new repository,
without any work done yet, creating a new folder and RRT files, or you can initiate a RRT
repository inside an existing project/folder you already have. If the latter, we don't alter
your files at all, but simply write a few files needed for RRT to work properly. By detault
repo initialization is done interactively, so that you can choose your settings or accept
reasonable defaults.
}
\details{
You can run this function to start a new RRT repo, and to refresh a repo with new work. This
function downloads packages and installs them, as needed.
}
\examples{
\dontrun{
# new repo
rrt(repo="~/scottsnewrepo")

# refresh repo
rrt(repo="~/scottsnewrepo")
}
}
\seealso{
\link{rrt_refresh}, \link{rrt_init}
}

