% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{rrt_bitbucket}
\alias{rrt_bitbucket}
\title{Create Bitbucket repository}
\usage{
rrt_bitbucket(name, description = "", private = FALSE,
  fork_policy = "allow_forks", has_issues = TRUE, has_wiki = TRUE,
  website = "", browse = TRUE, ...)
}
\arguments{
\item{name}{Required. Name of the repository}

\item{description}{A string containing the repository's description.}

\item{private}{A boolean indicating if a repository is private or public.}

\item{fork_policy}{Control the rules for forking this repository. See Details}

\item{has_issues}{A boolean indicating a repository has an issue tracker.}

\item{has_wiki}{A boolean indicating if the repository has a wiki.}

\item{website}{Any website to go in the repository metadata}

\item{browse}{(logical) If TRUE, at the end of the function the new repo is opened in your
default browser}

\item{...}{Further curl options passed on to httr::POST}
}
\description{
Create Bitbucket repository
}
\details{
Available values for fork_policy parameter are:
\itemize{
 \item allow_forks: unrestricted forking
 \item no_public_forks: restrict forking to private forks (forks cannot be made public later)
 \item no_forks: deny all forking
}
}
\examples{
\dontrun{
rrt_bitbucket(name="thingsandstuff", description="some comments")
}
}

