% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rrt_browse}
\alias{rrt_browse}
\title{Browse rrt libraries}
\usage{
rrt_browse(repoid = NULL, output = NULL)
}
\arguments{
\item{repoid}{Respository id, default is NULL, so gets all repos}

\item{output}{File to output RRT dashboard file to.}
}
\description{
Open up a dashboard in your default browse.
}
\examples{
rrt_browse()
}

