% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rrt_compat}
\alias{rrt_compat}
\title{Local test for package compatibility}
\usage{
rrt_compat(repo)
}
\arguments{
\item{repo}{Repository path}

\item{what}{What to test, one or more of install, tests, examples.}
}
\description{
Local test for package compatibility
}
\examples{
rrt_compat(repo="~/testrepo")
}

