% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{rrt_git}
\alias{rrt_git}
\title{Commit versions of an RRT repository to git}
\usage{
rrt_git(repo = NULL, verbose = TRUE)
}
\arguments{
\item{repo}{(character) A path to create a RRT repository; defaults to current working directory.}

\item{verbose}{(logical) Whether to print messages or not (Default: TRUE).}
}
\value{
Files written to the user system, with informative messages on progress
}
\description{
Commit versions of an RRT repository to git
}
\examples{
\dontrun{
rrt_git(repo="~/testrepo")
}
}
\keyword{internal}

