% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rrt_github}
\alias{rrt_github}
\title{Create a Github repository}
\usage{
rrt_github(name, description = "", homepage = "", private = FALSE,
  has_issues = TRUE, has_wiki = TRUE, has_downloads = TRUE,
  team_id = NULL, auto_init = FALSE, gitignore_template = NULL,
  license_template = NULL, browse = TRUE, ...)
}
\arguments{
\item{name}{(character) Required. The name of the repository}

\item{description}{(character)    A short description of the repository}

\item{homepage}{(character)    A URL with more information about the repository}

\item{private}{(boolean)    Either true to create a private repository, or false to create a public
one. Creating private repositories requires a paid GitHub account. Default: false}

\item{has_issues}{(boolean)    Either true to enable issues for this repository, false to disable
them. Default: true}

\item{has_wiki}{(boolean)    Either true to enable the wiki for this repository, false to disable
it. Default: true}

\item{has_downloads}{(boolean)    Either true to enable downloads for this repository, false to
disable them. Default: true}

\item{team_id}{(numberic)    The id of the team that will be granted access to this repository.
This is only valid when creating a repository in an organization.}

\item{auto_init}{(boolean)    Pass true to create an initial commit with empty README. Default:
false}

\item{gitignore_template}{(character)    Desired language or platform .gitignore template to
apply. Use the name of the template without the extension. For example, 'Haskell'. Ignored
if the auto_init parameter is not provided.}

\item{license_template}{(character)    Desired LICENSE template to apply. Use the name of the
template without the extension. For example, 'mit' or 'mozilla'. Ignored if the auto_init
parameter is not provided.}

\item{browse}{(logical) If TRUE, at the end of the function the new repo is opened in your
default browser}

\item{...}{Further curl options passed on to httr::POST}
}
\description{
Create a Github repository
}
\examples{
\dontrun{
rrt_github(name="yeppers_peppers", description='some comments')
}
}

