% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rrt_install2}
\alias{rrt_install2}
\title{Install packages within an RRT repository.}
\usage{
rrt_install2(repo = getwd(), repoid, lib, suggests = FALSE,
  verbose = TRUE)
}
\arguments{
\item{repo}{A repository path. This is the path to the root of your RRT repository. By default,
we use the current working directory via \code{getwd()}.}

\item{repoid}{Repository ID.}

\item{lib}{Library path}

\item{suggests}{Install suggests or not. Default: FALSE.}

\item{verbose}{Print messages. Default: TRUE.}
}
\description{
Install packages within an RRT repository.
}

