% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rrt_installed_pkgs}
\alias{rrt_installed_pkgs}
\title{Check what packages are installed.}
\usage{
rrt_installed_pkgs(repo = getwd(), simplify = TRUE)
}
\arguments{
\item{repo}{A repository path. This is the path to the root of your RRT repository. By default,
we use the current working directory via \code{getwd()}.}

\item{simplify}{(logical) Whether to simplify to just package name and version or print entire
matrix of package information.}
}
\description{
Check what packages are installed.
}
\examples{
\dontrun{
rrt_installed_pkgs()
rrt_installed_pkgs(simplify=FALSE)
}
}

