% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rrt_refresh}
\alias{rrt_refresh}
\title{Refresh package - look for any new packages used and install those in rrt library}
\usage{
rrt_refresh(repo = getwd(), verbose = TRUE)
}
\arguments{
\item{repo}{(character) A path to create a RRT repository; defaults to current working directory.}

\item{verbose}{(logical) Whether to print messages or not (Default: TRUE).}
}
\description{
Refresh package - look for any new packages used and install those in rrt library
}

