% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rrt_repos_list}
\alias{rrt_repos_list}
\title{Read rrt libraries from user's .Rprofile file}
\usage{
rrt_repos_list(repoid = NULL)
}
\arguments{
\item{repoid}{Respository id, default is NULL, so gets all repos}
}
\description{
Read rrt libraries from user's .Rprofile file
}
\examples{
(repos <- rrt_repos_list())
names(repos)
rrt_repos_list(names(repos)[1])
rrt_repos_list(names(repos)[2])
}

