% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rrt_repos_remove}
\alias{rrt_repos_remove}
\title{Remove an RRT repo from the user's global list of RRT repos}
\usage{
rrt_repos_remove(repo = getwd(), verbose = TRUE)
}
\arguments{
\item{repo}{(character) A path to create a RRT repository; defaults to current working directory}

\item{verbose}{Print messages or not, Default: TRUE}
}
\value{
Prints message saying what has been done.
}
\description{
Remove an RRT repo from the user's global list of RRT repos
}
\examples{
\dontrun{
rrt_repos_remove(repo="~/testrepo")
}
}

