% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rrt_repos_write}
\alias{rrt_repos_write}
\title{Write rrt libraries to user's .Rprofile file}
\usage{
rrt_repos_write(repo, repoid)
}
\arguments{
\item{repo}{Respository path}

\item{repoid}{Repository id}
}
\description{
Write rrt libraries to user's .Rprofile file
}
\examples{
rrt_repos_write(repo="~/testrepo/")
}
\keyword{internal}

