% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rrt_sweep}
\alias{rrt_sweep}
\title{Clean out installed packages in your repository}
\usage{
rrt_sweep(repo, verbose = TRUE)
}
\arguments{
\item{repo}{Repository path}

\item{verbose}{(logical) Print messages (default) or not}
}
\description{
Clean out installed packages in your repository
}
\examples{
\dontrun{
rrt_sweep(repo="~/testrepo")
}
}

