% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rrtPath}
\alias{rrtPath}
\title{Returns folder paths for rrt library and src for repo.}
\usage{
rrtPath(repo, type = c("lib", "src", "manifest"))
}
\arguments{
\item{repo}{Repo path}

\item{type}{Either "lib" or "src"}
}
\description{
Returns folder paths for rrt library and src for repo.
}
\seealso{
Other Repo path: \code{\link{createRepoFolders}}
}

