% Generated by roxygen2 (4.0.1): do not edit by hand
\name{scanRepoPackages}
\alias{scanRepoPackages}
\title{Detect repository packages.}
\usage{
scanRepoPackages(repo = getwd(), fileext = NULL, verbose = TRUE)
}
\arguments{
\item{repo}{Repository path. Defaults to current working directory.}

\item{fileext}{(character) File extensions to search for. Default is .R, .Rmd, .Rnw, and .Rpres.  Specify a certain file extension(s) in a charcter vector. Exclude a file extension by e.g. "-.Rmd"}

\item{verbose}{(logical) Print messages or not.}
}
\value{
Vector of package names on which R code in the repository depends.
}
\description{
Detect all packages used in repository. Parses all .R, .Rmd, .Rnw, .Rpres, and other files in the repository directory to determine what packages the repo depends directly.
}
\details{
Only direct dependencies are detected (i.e. no recursion).

Packages are determined by parsing repository source code and looking for calls to \code{library}, \code{require}, \code{::}, and \code{:::}.
}
\note{
Modified from dependencies.R in the packrat github repo
}
\examples{
\dontrun{
   # dependencies for the repo in the current working dir
   scanRepoPackages()
  
   # dependencies for an repo in another directory
   scanRepoPackages("~/newrepo")
  
   # include only certain file extensions
   scanRepoPackages(fileext=c('Rmd'))
  
   # exclude some file extensions
   scanRepoPackages(fileext=c('-Rmd'))
   scanRepoPackages(fileext=c('-Rmd','Rnw'))
  
   # suppress messages
   scanRepoPackages(verbose=FALSE)
}

}
\seealso{
Other rrt: \code{\link{checkpoint}};
  \code{\link{print.rrtrepos}},
  \code{\link{rrt_repos_list}},
  \code{\link{rrt_repos_write}}; \code{\link{rrt_browse}};
  \code{\link{rrt_compat}}; \code{\link{rrt_init}};
  \code{\link{rrt_installed_pkgs}};
  \code{\link{rrt_refresh}};
  \code{\link{rrt_repos_remove}}; \code{\link{rrt_sweep}}
}

