% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setLibPaths}
\alias{setLibPaths}
\title{Configure library path so that repo library is at top of search path.}
\usage{
setLibPaths(repo, libPath = rrtPath(repo, "lib"))
}
\arguments{
\item{repo}{A repository path. This is the path to the root of your RRT repository. Defaults to current working directory current working directory via \code{\link{getwd}}.}

\item{libPath}{(character) Location of package library in repo}
}
\description{
Configure library path so that repo library is at top of search path.
}

