% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setMranMirror}
\alias{setMranMirror}
\title{Sets CRAN mirror to MRAN snapshot.}
\usage{
setMranMirror(snapshotdate, autosnap = "last",
  snapshotUrl = getSnapShotUrl(snapshotdate, autosnap = autosnap))
}
\arguments{
\item{snapshotUrl}{URL for CRAN snapshot mirror}

\item{snapshotdate}{(date) Required. Date of snapshot to use. E.g. "2014-06-20". If left blank, you
will be supplied with options.}

\item{autosnap}{One of first, last or all. Determines how to resolve the snapshot id for a given date, if more than one snapshot exists on MRAN for that date.}
}
\description{
Sets CRAN mirror to MRAN snapshot.
}

