% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkpoint}
\alias{checkpoint}
\title{Initiate or refresh an RRT repository.}
\usage{
checkpoint(snapshotdate = NULL, repo = getwd(), mran = TRUE,
  autosnap = TRUE, rprofile = NULL, interactive = FALSE,
  suggests = FALSE, verbose = TRUE)
}
\arguments{
\item{snapshotdate}{(date) Required. Date of snapshot to use. E.g. "2014-06-20". If left blank, you
will be supplied with options.}

\item{repo}{A repository path. This is the path to the root of your RRT repository. Defaults to current working directory current working directory via \code{\link{getwd}}.}

\item{mran}{(logical) If TRUE (default), packages are installed from the MRAN server. See \url{http://mran.revolutionanalytics.com} for more information.}

\item{autosnap}{(logical) Get most recent snapshot. Default: FALSE}

\item{verbose}{(logical) Whether to print messages or not (Default: TRUE).}

\item{rprofile}{(list) pass in a list of options to include in the \code{.Rprofile} file for the repo.}

\item{interactive}{(logical) If TRUE, function asks you for input for each item, otherwise, defaults are used. Default: FALSE.}

\item{suggests}{(logical) Download and install packages in the Suggests line for packages used in your RRT repository, or not. Default: FALSE.}
}
\value{
Files written to the user's machine, with informative messages on progress
}
\description{
This function decides what to do with your repository based on commands you give, and on
skimming through your repo. You can run this function to start a new repository,
without any work done yet, creating a new folder and RRT files, or you can initiate a RRT
repository inside an existing project/folder you already have. If the latter, we don't alter
your files at all, but simply write a few files needed for RRT to work properly. By detault
repo initialization is done interactively, so that you can choose your settings or accept
reasonable defaults.
}
\details{
You can run this function to start a new RRT repo, and to refresh a repo with new work. This
function downloads packages and installs them, as needed.
}
\examples{
\dontrun{
# new repo, after entering a project folder, starting R, then loading RRT
checkpoint("2014-07-11")

# refresh repo
checkpoint()
}
}
\seealso{
\link{rrt_refresh}, \link{rrt_init}

Other rrt: \code{\link{print.rrtrepos}},
  \code{\link{rrt_repos_list}},
  \code{\link{rrt_repos_write}}; \code{\link{rrt_browse}};
  \code{\link{rrt_compat}}; \code{\link{rrt_init}};
  \code{\link{rrt_installed_pkgs}};
  \code{\link{rrt_refresh}};
  \code{\link{rrt_repos_remove}}; \code{\link{rrt_sweep}};
  \code{\link{scanRepoPackages}}
}

