% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkpoint}
\alias{checkpoint}
\title{Configures R session to use packages as they existed on CRAN at a snapshot date.}
\usage{
checkpoint(snapshotDate, project = getwd(), verbose = TRUE)
}
\arguments{
\item{snapshotDate}{Date of snapshot to use in \code{YYYY-MM-DD} format,e.g. \code{"2014-09-17"}.  Daily snapshots exists on the MRAN server starting from \code{"2014-09-17"}.}

\item{project}{A project path. This is the path to the root of the project you want checkpointed. Defaults to current working directory using \code{\link{getwd}()}.}

\item{verbose}{If TRUE, displays progress messages.}
}
\value{
NULL. See the \code{details} section for side effects.
}
\description{
Together, the RRT package and the MRAN server act as a CRAN time machine. The \code{checkpoint()} function downloads the packages to a local library exactly as they were at the specified point in time. Only those packages are available to your session, thereby avoiding any package updates that came later and may have altered your results. In this way, anyone using RRT's \code{checkpoint()} can ensure the reproducibility of your scripts or projects at any time.
}
\section{Details}{


\code{checkpoint()} creates a local library into which it installs a copy of the packages required by your project as they were on CRAN on the specified snapshot date. Your R session is updated to use only these packages.

To automatically determine all packages used in your project, the function scans all R code files for \code{library()} and \code{requires()} statements.  During this scan, the function searches for all R and Rmarkdown scripts, in other words file with extensions \code{.R}, \code{.Rmd} and \code{.Rpres}.

Specifically, the function will:

\itemize{
\item{Create a new local snapshot library to download packages. This library folder is at \code{~/.rrt}}
\item{Update the options for your CRAN mirror and point to an MRAN snapshot using \code{options(repos)}}
\item{Scan your project folder for all required packages and install them from the snapshot using \code{\link[utils]{download.packages}}}
}
}
\examples{
\dontrun{

# Create temporary project and set working directory

example_project <- paste0("~/rrt_example_project_", Sys.Date())

dir.create(example_project, recursive = TRUE)
oldwd <- setwd(example_project)


# Write dummy code file to project

cat("library(MASS)", "library(foreach)",
    sep="\\n", 
    file="rrt_example_code.R")


# Create a checkpoint by specifying a snapshot date

library(RRT)
checkpoint("2014-09-17")

# Check that CRAN mirror is set to MRAN snapshot
getOption("repos")

# Check that library path is set to ~/.rrt
.libPaths()

# Check which packages are installed in RRT library
installed.packages()

# cleanup
unlink(example_project, recursive = TRUE)
setwd(oldwd)

}

}

