% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpoint.R
\name{checkpoint}
\alias{checkpoint}
\title{Configures R session to use packages as they existed on CRAN at time of snapshot.}
\usage{
checkpoint(snapshotDate, project = getwd(), R.version,
  scanForPackages = TRUE, checkpointLocation = "~/", verbose = TRUE,
  use.knitr = system.file(package = "knitr") != "",
  auto.install.knitr = TRUE, scan.rnw.with.knitr = FALSE,
  forceInstall = FALSE, forceProject = FALSE)
}
\arguments{
\item{snapshotDate}{Date of snapshot to use in \code{YYYY-MM-DD} format,e.g. \code{"2014-09-17"}.  Specify a date on or after \code{"2014-09-17"}.  MRAN takes one snapshot per day.}

\item{project}{A project path.  This is the path to the root of the project that references the packages to be installed from the MRAN snapshot for the date specified for \code{snapshotDate}.  Defaults to current working directory using \code{[getwd()]}.}

\item{R.version}{Optional character string, e.g. "3.1.2".  If specified, compares the current \code{[base::R.version]} to the specified R.version. If these differ, stops processing with an error, making no changes to the system. Specifically, if the check fails, the library path is NOT modified. This argument allows the original script author to specify a specific version of R to obtain the desired results.}

\item{scanForPackages}{If TRUE, scans for packages in project folder (see details). If FALSE, skips the scanning process.  A use case for \code{scanForPackages = FALSE} is to skip the scanning and installation process, e.g. in production environments with a large number of R scripts in the project.  Only set \code{scanForPackages = FALSE} if you are certain that all package dependencies are already in the checkpoint folder.}

\item{checkpointLocation}{File path where the checkpoint library is stored.  Default is \code{"~/"}, i.e. the user's home directory. A use case for changing this is to create a checkpoint library on a portable drive (e.g. USB drive).}

\item{verbose}{If TRUE, displays progress messages.}

\item{use.knitr}{If TRUE,  parses all \code{Rmarkdown} files using the \code{knitr} package.}

\item{auto.install.knitr}{If TRUE and the project contains rmarkdown files, then automatically included the packages \code{knitr} and \code{rmarkdown} in packages to install.}

\item{scan.rnw.with.knitr}{If TRUE, uses \code{[knitr::knit()]} to parse \code{.Rnw} files, otherwise use \code{[utils::Sweave()]}}

\item{forceInstall}{If TRUE, forces the re-installation of all discovered packages and their dependencies. This is useful if, for some reason, the checkpoint archive becomes corrupted.}

\item{forceProject}{If TRUE, forces the checkpoint process, even if the provided project folder doesn't look like an R project. A commonly reported user problem is that they accidentally trigger the checkpoint process from their home folder, resulting in scanning many R files and downloading many packages. To prevent this, we use a heuristic to determine if the project folder looks like an R project. If the project folder is the home folder, and also contains no R files, then \code{checkpoint()} asks for confirmation to continue.}
}
\value{
Checkpoint is called for its side-effects (see the details section), but invisibly returns a list with elements:
\itemize{
\item files_not_scanned
\item pkgs_found
\item pkgs_not_on_mran
\item pkgs_installed
}
}
\description{
Together, the checkpoint package and the checkpoint server act as a CRAN time machine.  The \code{checkpoint()} function installs the packages referenced in the specified project to a local library exactly as they existed at the specified point in time.  Only those packages are available to your session, thereby avoiding any package updates that came later and may have altered your results.  In this way, anyone using the checkpoint \code{checkpoint()} function can ensure the reproducibility of your scripts or projects at any time.
}
\section{Details}{


\code{checkpoint()} creates a local library into which it installs a copy of the packages required by your project as they existed on CRAN on the specified snapshot date.  Your R session is updated to use only these packages.

To automatically determine all packages used in your project, the function scans all R code (\code{.R}, \code{.Rmd}, and \code{.Rpres} files) for \code{[library()] and}\code{\link[=require]{require()}}\code{statements. In addition, scans for occurrences of code that accesses functions in namespaces using}package\link[:]{::}foo()\code{and}package\link[:::]{::::}foo()\code{. Finally, any occurrences of the functions}\link[methods:setClass]{methods::setClass}\code{,}\link[methods:setRefClass]{methods::setRefClass}\code{,}[methods::setMethod\code{or}[methods::setGeneric]\code{will also identify the}[methods]` package as a dependency.

Specifically, the function will:
\itemize{
\item Create a new local snapshot library to install packages.  By default this library folder is at \code{~/.checkpoint} but you can modify the path using the \code{checkpointLocation} argument.
\item Update the options for your CRAN mirror and point to an MRAN snapshot using \code{[base::options](repos)}
\item Scan your project folder for all required packages and install them from the snapshot using \code{[utils::install.packages()]}
}

[methods::setMethod\code{or}[methods::setGeneric]: R:methods::setMethod%60%20or%20%60[methods::setGeneric
[methods]: R:methods
[base::options]: R:base::options
[utils::install.packages()]: R:utils::install.packages()
}

\section{Resetting the checkpoint}{

To reset the checkpoint, simply restart your R session.
}

\section{Changing the default MRAN url}{


\code{checkpoint} uses https by default to download packages (see \url{https://www.r-consortium.org/blog/2015/08/17/best-practices-for-using-r-securely}).
\code{checkpoint} Defaults to \url{https://mran.microsoft.com/snapshot} by default in R versions 3.2.0 and later, if https support is enabled.

You can modify the default URL. To change the URL, use \code{options(checkpoint.mranUrl = ...)}
}

\section{Log file}{


As a side effect, the \code{checkpoint} function writes a log file with information about the downloaded files, in particular the package downloaded and the associated file size in bytes. The log is stored at the root of the \code{checkpointLocation}. For example, if \code{checkpointLocation} is the user home folder (the default) then the log file is at \code{~/.checkpoint/checkpoint_log.csv}. This file contains columns for:
\itemize{
\item \code{timestamp}
\item \code{snapshotDate}
\item \code{pkg}
\item \code{bytes}
}
}

\examples{
\dontrun{

# Create temporary project and set working directory

example_project <- paste0("~/checkpoint_example_project_", Sys.Date())

dir.create(example_project, recursive = TRUE)
oldwd <- setwd(example_project)


# Write dummy code file to project

cat("library(MASS)", "library(foreach)",
    sep="\\n",
    file="checkpoint_example_code.R")


# Create a checkpoint by specifying a snapshot date

library(checkpoint)
checkpoint("2014-09-17")

# Check that CRAN mirror is set to MRAN snapshot
getOption("repos")

# Check that library path is set to ~/.checkpoint
.libPaths()

# Check which packages are installed in checkpoint library
installed.packages()

# cleanup
unlink(example_project, recursive = TRUE)
setwd(oldwd)

}

}
