% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getPkgs}
\alias{getPkgs}
\title{Function to install pkgs}
\usage{
getPkgs(x, lib, recursive = FALSE, verbose = TRUE, install = TRUE)
}
\arguments{
\item{x}{(character) A vector of package names. If NULL, none installed, and message prints}

\item{lib}{(character) Library location, a directory}

\item{recursive}{(logical) Recursively install packages?}

\item{verbose}{(logical) Inherited from call to rrt_init or rrt_refresh}
}
\description{
Function to install pkgs
}
\examples{
\dontrun{
getPkgs()
}
}

