% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getsysreq}
\alias{getsysreq}
\title{Function to get system requireqments, if any, from installed packages}
\usage{
getsysreq(x)
}
\arguments{
\item{x}{(character) Name of package. One to many in a vector or list}
}
\description{
Function to get system requireqments, if any, from installed packages
}
\examples{
getsysreq('RCurl')
getsysreq(c('RCurl','doMC','ggplot2','XML','rgdal'))
}
\keyword{internal}

