% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSnapshot.R
\name{getValidSnapshots}
\alias{getValidSnapshots}
\title{Read list of available snapshot dates from MRAN.}
\usage{
getValidSnapshots(mranRootUrl = mranUrl())
}
\arguments{
\item{mranRootUrl}{MRAN root. This can be a URL, e.g. \code{https://mran.microsoft.com/snapshot/} or the path to a local MRAN repository, e.g.\code{file:///local/path}}
}
\value{
Character vector with dates of valid snapshots
}
\description{
Returns vector of available dates from MRAN or local MRAN repository.
}
\seealso{
Other checkpoint functions: \code{\link{checkpointArchives}},
  \code{\link{checkpointRemove}}, \code{\link{checkpoint}},
  \code{\link{getAccessDate}}, \code{\link{mranUrl}},
  \code{\link{setSnapshot}}, \code{\link{unCheckpoint}}
}
