% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mranUrl.R
\name{getValidSnapshots}
\alias{getValidSnapshots}
\title{Read list of available snapshot dates from MRAN url.}
\usage{
getValidSnapshots(mranRootUrl = mranUrl())
}
\arguments{
\item{mranRootUrl}{URL of MRAN root, e.g. \code{"https://mran.microsoft.com/snapshot/"} or \code{"file:///local/path"}}
}
\description{
Read list of available snapshot dates from MRAN url.
}

