% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{git_add_commit}
\alias{git_add_commit}
\title{Add and commit files in your RRT repository using git.}
\usage{
git_add_commit(path = getwd(), message = "", verbose = TRUE, ...)
}
\arguments{
\item{path}{Path to the RRT repository}

\item{message}{Commit message. If you don't pass a message we ask you for a commit message.
Use quotes around your commit message!}

\item{verbose}{Print messages or not, Default: TRUE}

\item{...}{Further args passed on to \code{git2r::commit}}
}
\description{
Add and commit files in your RRT repository using git.
}
\details{
Note this doesn't do anything about remote connections to your repo. See help for
\link{remote_add}
}
\examples{
\dontrun{
git_add_commit(path="~/bbb", message="Made some changes to the regression function")
}
}

