% Generated by roxygen2 (4.0.1): do not edit by hand
\name{install_mran}
\alias{install_mran}
\title{Install packages from MRAN}
\usage{
install_mran(pkg, repo, snapshotdate = NULL, libPath = rrtPath(repo, "lib"),
  srcPath = rrtPath(repo, "src"), snapshotid = getSnapshotId(snapshotdate),
  quiet = FALSE, ..., dependencies = TRUE)
}
\arguments{
\item{pkg}{Package name}

\item{...}{Further args passed on to \code{\link[utils]{install.packages}}}

\item{dependencies}{Whether to install dependencies or not. Default: TRUE. See}

\item{repo}{A repository path. This is the path to the root of your RRT repository. Defaults to current working directory current working directory via \code{\link{getwd}}.}

\item{snapshotdate}{(date) Required. Date of snapshot to use. E.g. "2014-06-20". If left blank, you
will be supplied with options.}

\item{quiet}{Passed to \code{\link[utils]{install.packages}}}

\item{libPath}{(character) Location of package library in repo}

\item{srcPath}{(character) Location of package src in repo}

\item{snapshotid}{(character) The identifying snapshot id, consisting of a date and semi-timestamp, e.g. "2014-08-01_0500"}
}
\value{
Installs a package, or throws warnings/errors on failures
}
\description{
Steps (for each pkg):
 download package
 determine dependencies
 download dependencies (but don't install them)
 install target pkg and downloaded dependencies
}
\examples{
\dontrun{
  repo_path <- tempdir()
  createRepoFolders(repo_path)
  
  install_mran(repo=repo_path, pkg="plyr", snapshotdate="2014-08-01")
  
  dir(instPath)
  unlink(instPath, recursive = TRUE)
}



}
\seealso{
Other mran: \code{\link{mranPkgMetadata}};
  \code{\link{mranSnapshots}}; \code{\link{mran_diffs}};
  \code{\link{pkgVersionAtSnapshot}}
}

