% Generated by roxygen2 (4.0.1): do not edit by hand
\name{install_mran}
\alias{install_mran}
\title{Install packages from MRAN}
\usage{
install_mran(pkg, date = NULL, lib = NULL, destdir = NULL,
  quiet = FALSE, ..., dependencies = TRUE)
}
\arguments{
\item{pkg}{Package name}

\item{date}{MRAN snapshot date}

\item{lib}{Library to install packages in}

\item{destdir}{Destination directory to install packages to}

\item{...}{Further args passed on to \code{\link[utils]{install.packages}}}

\item{dependencies}{Whether to install dependencies or not. Default: TRUE. See}

\item{quiet}{Passed to \code{\link[utils]{install.packages}}
\link{install.packages}}
}
\value{
Installs a package, or throws warnings/errors on failures
}
\description{
Steps (for each pkg):
 download package
 determine dependencies
 download dependencies (but don't install them)
 install target pkg and downloaded dependencies
}
\examples{
\dontrun{
install_mran(pkg="plyr", date="2014-08-01", lib="~/mran_snaps/")
install_mran(pkg="calmate", date="2014-08-01", lib="~/mran_snaps/")
}
}
\seealso{
Other mran: \code{\link{mran_diffs}};
  \code{\link{mran_pkg_metadata}};
  \code{\link{mran_pkg_versions}}; \code{\link{mran_set}};
  \code{\link{mran_snaps}}
}

