% Generated by roxygen2 (4.0.1): do not edit by hand
\name{install_mran}
\alias{install_mran}
\title{Install packages from MRAN}
\usage{
install_mran(pkg, date = NULL, lib = NULL, destdir = NULL, ...,
  dependencies = TRUE)
}
\arguments{
\item{pkg}{Package name}

\item{date}{MRAN snapshot date}

\item{lib}{Library to install packages in}

\item{destdir}{Destination directory to install packages to}

\item{...}{Further args passed on to \link{install.packages}}

\item{dependencies}{Whether to install dependencies or not. Default: TRUE. See
\link{install.packages}}
}
\value{
Installs a package, or throws warnings/errors on failures
}
\description{
Steps (for each pkg):
 download package
 determine dependencies
 download dependencies (but don't install them)
 install target pkg and downloaded dependencies
}
\examples{
\dontrun{
install_mran(pkg="plyr", date="2014-06-23", lib="~/mran_snaps/")
install_mran(pkg="calmate", date="2014-06-23", lib="~/mran_snaps/")
}
}

