% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is_rrt}
\alias{is_rrt}
\title{Find out if a repository is an RRT repository.}
\usage{
is_rrt(repo)
}
\arguments{
\item{repo}{A repository path. This is the path to the root of your RRT repository. Defaults to current working directory current working directory via \code{\link{getwd}}.}
}
\value{
A logical, TRUE or FALSE
}
\description{
Find out if a repository is an RRT repository.
}

