% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mran_diffs}
\alias{mran_diffs}
\title{Get available diffs from MRAN}
\usage{
mran_diffs(diff = NULL)
}
\arguments{
\item{diff}{Optional. (character) A diff date-time stamp of a MRAN diff.}
}
\description{
Get available diffs from MRAN
}
\examples{
\dontrun{
mran_diffs()

# An individual diff
mran_diffs(diff="2014-06-19_0136")

diffs <- mran_diffs()
mran_diffs(diffs[length(diffs)-1])
}
}

