% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mran_pkg_metadata}
\alias{mran_pkg_metadata}
\title{Get available package level metadata from Marmoset}
\usage{
mran_pkg_metadata(snapshot = NULL, package)
}
\arguments{
\item{snapshot}{A Marmoset snapshot. Defaults to most recent snapshot}

\item{package}{Required. A package name}
}
\description{
Get available package level metadata from Marmoset
}
\examples{
\dontrun{
mran_pkg_metadata(package="plyr")
}
}

