% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mran_pkg_metadata}
\alias{mran_pkg_metadata}
\title{Get available package level metadata from MRAN}
\usage{
mran_pkg_metadata(package, snapshot = NULL)
}
\arguments{
\item{package}{Required. A package name}

\item{snapshot}{An MRAN snapshot ('YYYY-MM-DD_TTTT') or a date ('YYYY-MM-DD'). Defaults to most
recent snapshot.}
}
\description{
Get available package level metadata from MRAN
}
\examples{
\dontrun{
mran_pkg_metadata(package="plyr", snapshot="2014-08-04")
}
}
\seealso{
Other mran: \code{\link{install_mran}};
  \code{\link{mran_diffs}};
  \code{\link{mran_pkg_versions}}; \code{\link{mran_set}};
  \code{\link{mran_snaps}}
}

