% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{mran_pkg_metadata}
\alias{mran_pkg_metadata}
\title{Get available package level metadata from MRAN}
\usage{
mran_pkg_metadata(package, snapshot = NULL)
}
\arguments{
\item{package}{Required. A package name}

\item{snapshot}{A MRAN snapshot. Defaults to most recent snapshot}
}
\description{
Get available package level metadata from MRAN
}
\examples{
\dontrun{
mran_pkg_metadata(package="plyr")
}
}

