% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mran_pkg_versions}
\alias{mran_pkg_versions}
\title{Get available package versions from MRAN}
\usage{
mran_pkg_versions(package, snapshot = NULL, which = "src", os = "macosx")
}
\arguments{
\item{package}{Required. A package name}

\item{snapshot}{An MRAN snapshot ('YYYY-MM-DD_TTTT') or a date ('YYYY-MM-DD'). Defaults to most
recent snapshot.}

\item{which}{one of src or bin}
}
\description{
Get available package versions from MRAN
}
\examples{
\dontrun{
mran_pkg_versions(snapshot="2014-07-14", package="plyr")
mran_pkg_versions(snapshot="2014-08-04", package="plyr", which="bin", os="windows")
}
}

