% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mranSnapshots}
\alias{mranSnapshots}
\title{Get available snapshots from MRAN}
\usage{
mranSnapshots(date = NULL, message = TRUE)
}
\arguments{
\item{date}{(character) A date, in the format YYY-MM-DD}
}
\description{
Get available snapshots from MRAN
}
\examples{
\dontrun{
# List all available snapshots
mranSnapshots()
# Get code for a single snapshot
mranSnapshots(date='2014-08-04')
}
}
\seealso{
Other mran: \code{\link{install_mran}};
  \code{\link{mranPkgMetadata}};
  \code{\link{mranPkgVersions}}; \code{\link{mran_diffs}}
}

