% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pkgVersionAtSnapshot}
\alias{pkgVersionAtSnapshot}
\title{Get available package versions from MRAN}
\usage{
pkgVersionAtSnapshot(pkgs, snapshotdate,
  snapshotid = snapshotFromDate(snapshotdate), type = c("src", "mac.binary",
  "win.binary"), Rversion = "R3.0")
}
\arguments{
\item{package}{(character) Required. A package name}

\item{snapshot}{(date) An MRAN snapshot ('YYYY-MM-DD_TTTT') or a date ('YYYY-MM-DD'). Defaults to most recent snapshot.}

\item{type}{(character) "src", "mac.binary" or "win.binary"}
}
\description{
Get available package versions from MRAN
}
\examples{
pkgVersionAtSnapshot(snapshot="2014-07-14", package="plyr")
pkgVersionAtSnapshot(snapshot="2014-08-01", package="plyr", type="win.binary")
}
\seealso{
Other mran: \code{\link{install_mran}};
  \code{\link{mranPkgMetadata}};
  \code{\link{mranSnapshots}}; \code{\link{mran_diffs}}
}

