% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repoDependencies}
\alias{repoDependencies}
\title{Get repository dependencies}
\usage{
repoDependencies(pkgs, simplify = FALSE, base = TRUE, suggests = FALSE,
  ...)
}
\arguments{
\item{repo}{(character) A path to a RRT repository; defaults to current working directory.}

\item{simplify}{(logical) If TRUE, simplify list to a vector with all unique packages.}

\item{base}{(logical) If TRUE, return base R packages, if FALSE, don't return them.}

\item{suggests}{(logical) Download and install packages in the Suggests line for packages
used in your RRT repository, or not. Default: FALSE.}

\item{...}{Further args passed on to \code{miniCRAN::pkgDep}}
}
\value{
A named list of packages, named by the package that requires said dependencies
}
\description{
This function first determines packages used in the repo, then recursively gets dependencies of the packages you use (for Depends, Imports and LinkLibrary only, see \code{?pkgDep}).  Optionally, you can get dependencies for Suggests and Enhances (non-recursively). NOTE: Enhances not working right now.
}
\keyword{internal}

