% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repodeps}
\alias{repodeps}
\title{Get repository dependencies}
\usage{
repodeps(repo = getwd(), simplify = FALSE)
}
\arguments{
\item{repo}{A path to a RRT repository; defaults to current working directory.}
}
\value{
A named list of packages, named by the package that requires said dependencies
}
\description{
This function first determines packages used in your repo, then recursively gets dependencies
of the packages you use (for Depends, Imports and LinkLibrary only, see \code{?pkgDep}).
}
\examples{
repodeps(repo="~/myrepo")
repodeps(repo="~/myrepo", simplify=TRUE)
setwd("~/myrepo")
repodeps()
}

