% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{package}
\name{RRT-package}
\alias{RRT}
\alias{RRT-package}
\title{R Reproducibility Toolkit}
\description{
R Reproducibility Toolkit
}
\details{
For help see the package vignette by navigating to vignettes in RRT help \code{?RRT}

An RRT repository is a like any normal folder on your computer, except we install a new folder
inside your repository that includes a manifest file to keep track of all the packages you
use in the repository, as well as the packages locally installed.
}
\keyword{package}

