% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rrt_browse}
\alias{rrt_browse}
\title{Browse rrt libraries}
\usage{
rrt_browse(repoid = NULL, output = NULL, browse = TRUE)
}
\arguments{
\item{repoid}{Respository id, default is NULL, so gets all repos}

\item{output}{File to output RRT dashboard file to.}

\item{browse}{(logical) If TRUE (default), web page opens in your default browse. If FALSE, html
file written to disk and path to that file printed to console.}
}
\description{
Open up a dashboard in your default browse.
}
\examples{
\dontrun{
rrt_browse()
}
}
\seealso{
Other rrt: \code{\link{checkpoint}};
  \code{\link{print.rrtrepos}},
  \code{\link{rrt_repos_list}},
  \code{\link{rrt_repos_write}}; \code{\link{rrt_compat}};
  \code{\link{rrt_init}}; \code{\link{rrt_installed_pkgs}};
  \code{\link{rrt_packages}}; \code{\link{rrt_refresh}};
  \code{\link{rrt_repos_remove}}; \code{\link{rrt_sweep}}
}

