% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rrt_github}
\alias{rrt_bitbucket}
\alias{rrt_github}
\title{Create a Github repo}
\usage{
rrt_github(name, description = "", homepage = "", private = FALSE,
  has_issues = TRUE, has_wiki = TRUE, has_downloads = TRUE,
  team_id = NULL, auto_init = FALSE, gitignore_template = NULL,
  license_template = NULL, verbose = TRUE, browse = TRUE, ...)

rrt_bitbucket(name, description = "", private = FALSE,
  fork_policy = "allow_forks", has_issues = TRUE, has_wiki = TRUE,
  website = "", verbose = TRUE, browse = TRUE, ...)
}
\arguments{
\item{name}{(character) Required. The name of the repository}

\item{description}{(character)    A short description of the repository}

\item{homepage}{(character)    A URL with more information about the repository}

\item{private}{(boolean)    Either true to create a private repository, or false to create a public
one. Creating private repositories requires a paid GitHub account. Default: false}

\item{has_issues}{(boolean)    Either true to enable issues for this repository, false to disable
them. Default: true}

\item{has_wiki}{(boolean)    Either true to enable the wiki for this repository, false to disable
it. Default: true}

\item{has_downloads}{(boolean)    Either true to enable downloads for this repository, false to
disable them. Default: true}

\item{team_id}{(numberic)    The id of the team that will be granted access to this repository.
This is only valid when creating a repository in an organization.}

\item{auto_init}{(boolean)    Pass true to create an initial commit with empty README. Default:
false}

\item{gitignore_template}{(character)    Desired language or platform .gitignore template to
apply. Use the name of the template without the extension. For example, “Haskell”. Ignored
if the auto_init parameter is not provided.}

\item{license_template}{(character)    Desired LICENSE template to apply. Use the name of the
template without the extension. For example, “mit” or “mozilla”. Ignored if the auto_init
parameter is not provided.}

\item{scm}{The source control manager for the repository. This is either hg or git.}

\item{has_wiki}{A boolean indicating if the repository has a wiki.}

\item{description}{A string containing the repository's description.}

\item{links}{An array of related objects.}

\item{updated_on}{A date timestamp of the last update to this repository.}

\item{fork_policy}{Control the rules for forking this repository. See Details}

\item{created_on}{An ISO-8601 date timestamp of this repository's creation date.}

\item{owner}{The owner's account.}

\item{size}{The size of the repository in bytes.}

\item{parent}{The parent repository this repository was forked off (only present on forks). This is a repository object itself.}

\item{has_issues}{A boolean indicating a repository has an issue tracker.}

\item{is_private}{A boolean indicating if a repository is private or public.}

\item{full_name}{The unique key into the repository. This key has the format: {owner}/{repo_slug}}

\item{name}{The display name of the repository.}

\item{language}{The main (programming) language of the repository source files.}
}
\description{
Create a Github repo

Create Bitbucket repository
}
\details{
Available values for fork_policy parameter are:
\itemize{
\item allow_forks: unrestricted forking
\item no_public_forks: restrict forking to private forks (forks cannot be made public later)
\item no_forks: deny all forking
}
}
\examples{
\dontrun{
rrt_github(name="yeppers_peppers", description='some comments')
rrt_bitbucket(name="thingsandstuff", description="some comments")
}
}

