% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rrt_init}
\alias{rrt_init}
\title{Initiate a RRT repository.}
\usage{
rrt_init(repo = getwd(), verbose = TRUE, rprofile = NULL,
  interactive = FALSE)
}
\arguments{
\item{repo}{(character) A path to create a RRT repository; defaults to current working directory.}

\item{verbose}{(logical) Whether to print messages or not (Default: TRUE).}

\item{rprofile}{(list) pass in a list of options to include in the .Rprofile file for the repo.}

\item{interactive}{(logical) If TRUE (default), function asks you for input for each item,
otherwise, defaults are used.}
}
\value{
Files written to the user system, with informative messages on progress
}
\description{
This function initiates a repository.
}
\examples{
\dontrun{
rrt_init(repo="~/testrepo")
rrt_refresh(repo="~/testrepo")
rrt_install(repo="~/testrepo")

# Optionally, do an interactive repo intitialization
rrt_init(repo="~/mynewcoolrepo", interactive=TRUE)
}
}

