% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rrt_init}
\alias{rrt_init}
\title{Initiate a RRT repository.}
\usage{
rrt_init(repo = getwd(), mran = FALSE, snapdate = NULL, verbose = TRUE,
  rprofile = NULL, interactive = TRUE)
}
\arguments{
\item{repo}{(character) A path to create a RRT repository; defaults to current working directory.}

\item{mran}{(logical) If TRUE, packages are installed from the MRAN server. See
\url{http://marmoset.revolutionanalytics.com/} for more information.}

\item{snapdate}{Date of snapshot to use. E.g. "2014-06-20"}

\item{verbose}{(logical) Whether to print messages or not (Default: TRUE).}

\item{rprofile}{(list) pass in a list of options to include in the .Rprofile file for the repo.}

\item{interactive}{(logical) If TRUE (default), function asks you for input for each item,
otherwise, defaults are used.}
}
\value{
Files written to the user's machine, with informative messages on progress
}
\description{
This function initiates a repository. You can run this function to start a new repository,
without any work done yet, creating a new folder and RRT files, or you can initiate a RRT
repository inside an existing project/folder you already have. If the latter, we don't alter
your files at all, but simply write a few files needed for RRT to work properly. By detault
repo initialization is done interactively, so that you can choose your settings or accept
reasonable defaults.
}
\examples{
\dontrun{
rrt_init(repo="~/testrepo")
rrt_refresh(repo="~/testrepo")
rrt_refresh(repo="~/testrepo", mran=TRUE)
rrt_install(repo="~/testrepo")

# Optionally, do an interactive repo intitialization
rrt_init(repo="~/mynewcoolrepo", interactive=TRUE)
}
}

