% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rrt_init}
\alias{rrt_init}
\title{Initiate a RRT repository.}
\usage{
rrt_init(repo = getwd(), mran = TRUE, snapdate = NULL, autosnap = FALSE,
  verbose = TRUE, rprofile = NULL, interactive = FALSE,
  suggests = FALSE, quiet = FALSE)
}
\arguments{
\item{repo}{(character) A path to create a RRT repository; defaults to current working directory.}

\item{mran}{(logical) If TRUE (default), packages are installed from the MRAN server. See
\url{http://mran.revolutionanalytics.com} for more information.}

\item{snapdate}{Date of snapshot to use. E.g. "2014-06-20"}

\item{autosnap}{(logical) Get most recent snapshot. Default: FALSE}

\item{verbose}{(logical) Whether to print messages or not (Default: TRUE).}

\item{rprofile}{(list) pass in a list of options to include in the .Rprofile file for the repo.}

\item{interactive}{(logical) If TRUE, function asks you for input for each item, otherwise, defaults are used. Default: FALSE.}

\item{suggests}{(logical) Download and install packages in the Suggests line for packages used in your RRT repository, or not. Default: FALSE.}

\item{quiet}{Passed to \code{\link[utils]{install.packages}}}
}
\value{
Files written to the user's machine, with informative messages on progress
}
\description{
This function initiates a repository. You can run this function to start a new repository, without any work done yet, creating a new folder and RRT files, or you can initiate a RRT repository inside an existing project/folder you already have. If the latter, we don't alter your files at all, but simply write a few files needed for RRT to work properly. By detault repo initialization is done interactively, so that you can choose your settings or accept reasonable defaults.
}
\examples{
\dontrun{
rrt_init(repo="~/testrepo")
rrt_refresh(repo="~/testrepo")
rrt_refresh(repo="~/testrepo", mran=TRUE)
rrt_install(repo="~/testrepo")

# Optionally, do an interactive repo intitialization
rrt_init(repo="~/mynewcoolrepo", interactive=TRUE)
}
}
\seealso{
\link{rrt_refresh}, \link{rrt_install}

Other rrt: \code{\link{checkpoint}};
  \code{\link{print.rrtrepos}},
  \code{\link{rrt_repos_list}},
  \code{\link{rrt_repos_write}}; \code{\link{rrt_browse}};
  \code{\link{rrt_compat}}; \code{\link{rrt_deps}};
  \code{\link{rrt_installed_pkgs}};
  \code{\link{rrt_refresh}};
  \code{\link{rrt_repos_remove}}; \code{\link{rrt_sweep}}
}

