% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rrt_install}
\alias{rrt_install}
\title{Install packages}
\usage{
rrt_install(repo, verbose = TRUE)
}
\arguments{
\item{repo}{A repository path}

\item{verbose}{Print messages
rrt_install(repo="~/testrepo")}
}
\description{
Install packages
}

