% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rrt_install}
\alias{rrt_install}
\title{Install packages within an RRT repository.}
\usage{
rrt_install(repo = getwd(), snapshot, libPath = rrtPath(repo, "lib"),
  srcPath = rrtPath(repo, "src"), suggests = FALSE, verbose = TRUE,
  quiet = FALSE)
}
\arguments{
\item{repo}{A repository path. This is the path to the root of your RRT repository. By default, we use the current working directory via \code{\link{getwd}}.}

\item{lib}{Library path}

\item{suggests}{Install suggests or not. Default: FALSE.}

\item{verbose}{Print messages. Default: TRUE}

\item{quiet}{Passed to \code{\link[utils]{install.packages}}}
}
\description{
Install packages within an RRT repository.
}
\keyword{internal}

