% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rrt_refresh}
\alias{rrt_refresh}
\title{Refresh package - look for any new packages used and install those in rrt library.}
\usage{
rrt_refresh(repo, snapshotdate = NULL, mran = TRUE, autosnap = TRUE,
  verbose = TRUE, suggests = FALSE, quiet = FALSE)
}
\arguments{
\item{repo}{A repository path. This is the path to the root of your RRT repository. Defaults to current working directory current working directory via \code{\link{getwd}}.}

\item{snapshotdate}{(date) Required. Date of snapshot to use. E.g. "2014-06-20". If left blank, you
will be supplied with options.}

\item{mran}{(logical) If TRUE (default), packages are installed from the MRAN server. See \url{http://mran.revolutionanalytics.com} for more information.}

\item{autosnap}{(logical) Get most recent snapshot. Default: FALSE}

\item{verbose}{(logical) Whether to print messages or not (Default: TRUE).}

\item{suggests}{(logical) Download and install packages in the Suggests line for packages used in your RRT repository, or not. Default: FALSE.}

\item{quiet}{Passed to \code{\link[utils]{install.packages}}}
}
\description{
Refresh package - look for any new packages used and install those in rrt library.
}
\seealso{
\code{\link{rrt_init}}, \code{\link{rrt_install}}

Other rrt: \code{\link{checkpoint}};
  \code{\link{print.rrtrepos}},
  \code{\link{rrt_repos_list}},
  \code{\link{rrt_repos_write}}; \code{\link{rrt_browse}};
  \code{\link{rrt_init}}; \code{\link{rrt_installed_pkgs}};
  \code{\link{rrt_repos_remove}}; \code{\link{rrt_sweep}};
  \code{\link{scanRepoPackages}}
}

