% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rrt_refresh}
\alias{rrt_refresh}
\title{Refresh package - look for any new packages used and install those in rrt library}
\usage{
rrt_refresh(repo = getwd(), mran = FALSE, snapdate = NULL,
  verbose = TRUE)
}
\arguments{
\item{repo}{(character) A path to create a RRT repository; defaults to current working directory.}

\item{mran}{(logical) If TRUE, packages are installed from the MRAN server. See
\url{http://marmoset.revolutionanalytics.com/} for more information.}

\item{snapdate}{Date of snapshot to use. E.g. "2014-06-20"}

\item{verbose}{(logical) Whether to print messages or not (Default: TRUE).}
}
\description{
Refresh package - look for any new packages used and install those in rrt library
}

