% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rrt_refresh}
\alias{rrt_refresh}
\title{Refresh package - look for any new packages used and install those in rrt library.}
\usage{
rrt_refresh(repo = getwd(), mran = TRUE, snapdate = NULL,
  autosnap = FALSE, verbose = TRUE, suggests = FALSE, quiet = FALSE)
}
\arguments{
\item{repo}{(character) A path to create a RRT repository; defaults to current working directory.}

\item{mran}{(logical) If TRUE (default), packages are installed from the MRAN server. See
\url{http://mran.revolutionanalytics.com} for more information.}

\item{snapdate}{Date of snapshot to use. E.g. "2014-06-20"}

\item{autosnap}{(logical) Get most recent snapshot. Default: FALSE}

\item{verbose}{(logical) Whether to print messages or not (Default: TRUE).}

\item{suggests}{(logical) Download and install packages in the Suggests line for packages used in your RRT repository, or not. Default: FALSE.}

\item{quiet}{Passed to \code{\link[utils]{install.packages}}}
}
\description{
Refresh package - look for any new packages used and install those in rrt library.
}
\examples{
\dontrun{
rrt_init(repo="~/testrepo")
rrt_refresh(repo="~/testrepo")
rrt_refresh(repo="~/testrepo", mran=TRUE)
rrt_install(repo="~/testrepo")

# Optionally, do an interactive repo intitialization
rrt_init(repo="~/mynewcoolrepo", interactive=TRUE)
}
}
\seealso{
\code{\link{rrt_init}}, \code{\link{rrt_install}}
}

