% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rrt_repos_remove}
\alias{rrt_repos_remove}
\title{Remove any missing RRT repos from the user's global list of RRT repos}
\usage{
rrt_repos_remove(repo = getwd(), verbose = TRUE)
}
\arguments{
\item{repo}{(character) A path to create a RRT repository; defaults to current working directory}

\item{verbose}{Print messages or not, Default: TRUE}
}
\value{
Prints message saying what has been done.
}
\description{
Remove any missing RRT repos from the user's global list of RRT repos
}
\seealso{
Other rrt: \code{\link{checkpoint}};
  \code{\link{print.rrtrepos}},
  \code{\link{rrt_repos_list}},
  \code{\link{rrt_repos_write}}; \code{\link{rrt_browse}};
  \code{\link{rrt_compat}}; \code{\link{rrt_init}};
  \code{\link{rrt_installed_pkgs}};
  \code{\link{rrt_refresh}}; \code{\link{rrt_sweep}};
  \code{\link{scanRepoPackages}}
}

