% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rrt_share}
\alias{rrt_share}
\title{Share an RRT repository}
\usage{
rrt_share(repo = getwd(), to = "zip", output = NULL, address = NULL,
  ...)
}
\arguments{
\item{repo}{An RRT repository}

\item{to}{How to share. One of zip, tar, github, gist, bitbucket, or email.}

\item{output}{Output. This can be one of a file name, directory, or X.}

\item{...}{Further args passed on to each internal method. See the examples and Details.}
}
\description{
Share an RRT repository
}
\details{
Note that with \code{to='email'} we can't attach the zip file for you, but we print out where
the attachment is so you can easily find it and attach it to the email quickly.
}
\examples{
\dontrun{
rrt_share(to='zip', output='~/myrepo.zip')
rrt_share(to='tar', output='~/myrepo.tar.gz')
rrt_share(to='gist')
rrt_share(to='github')
rrt_share(to='email')
}
}

