% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{rrt_sweep}
\alias{rrt_sweep}
\title{Clean out installed packages and their sources in your repository}
\usage{
rrt_sweep(repo = getwd(), pkgs = NULL, verbose = TRUE)
}
\arguments{
\item{repo}{Repository path, Defaults to the current working directory}

\item{pkgs}{A vector of package names}

\item{verbose}{(logical) Print messages (default) or not}
}
\description{
Clean out installed packages and their sources in your repository
}
\examples{
\dontrun{
rrt_refresh(repo="~/testrepo")
rrt_sweep(repo="~/testrepo")
}
}

