% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rrt_sweep}
\alias{rrt_sweep}
\title{Clean out installed packages and their sources in your repository}
\usage{
rrt_sweep(repo = getwd(), pkgs = NULL, verbose = TRUE)
}
\arguments{
\item{repo}{Repository path, Defaults to the current working directory}

\item{pkgs}{A vector of package names}

\item{verbose}{(logical) Print messages (default) or not}
}
\description{
Clean out installed packages and their sources in your repository
}
\examples{
\dontrun{
rrt_refresh(repo="~/testrepo")
rrt_sweep(repo="~/testrepo")
}
}
\seealso{
Other rrt: \code{\link{checkpoint}};
  \code{\link{print.rrtrepos}},
  \code{\link{rrt_repos_list}},
  \code{\link{rrt_repos_write}}; \code{\link{rrt_browse}};
  \code{\link{rrt_compat}}; \code{\link{rrt_init}};
  \code{\link{rrt_installed_pkgs}};
  \code{\link{rrt_refresh}};
  \code{\link{rrt_repos_remove}};
  \code{\link{scanRepoPackages}}
}

