% Generated by roxygen2 (4.0.1): do not edit by hand
\name{writeManifest}
\alias{writeManifest}
\title{Function to write manifest file}
\usage{
writeManifest(repo, libPath = rrtPath(repo, "lib"), packs, repoid,
  reponame = "", author = "", license = "", description = "",
  remote = "", snapshot = NULL, verbose = FALSE)
}
\arguments{
\item{repo}{repo root path}

\item{libPath}{libPathy to install packages in}

\item{packs}{Packages used in the repo}

\item{repoid}{Respository ID}

\item{reponame}{repo name}

\item{author}{Authors, comma separated}

\item{license}{License, e.g., MIT}

\item{description}{Character description of repo}

\item{remote}{Remote git/svn/mercurial repo}
}
\value{
Writes a RRT manifest file ("rrt_manifest.yml") to disk
}
\description{
Function to write manifest file
}
\keyword{internal}

