% Generated by roxygen2 (4.0.1): do not edit by hand
\name{writeManifest}
\alias{writeManifest}
\title{Function to write manifest file}
\usage{
writeManifest(repository, librar, packs, repoid)
}
\value{
Writes a RRT manifest file to disc
}
\description{
Function to write manifest file
}
\keyword{internal}

