% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{writeManifest}
\alias{writeManifest}
\title{Function to write manifest file}
\usage{
writeManifest(repository, librar, packs, repoid, reponame = "", author = "",
  license = "", description = "", remote = "", snapshot = NULL)
}
\arguments{
\item{repository}{Repository root path}

\item{librar}{Library to install packages in}

\item{packs}{Packages used in the repository}

\item{repoid}{Respository ID}

\item{reponame}{Repository name}

\item{author}{Authors, comma separated}

\item{license}{License, e.g., MIT}

\item{description}{Character description of repository}

\item{remote}{Remote git/svn/mercurial repository}
}
\value{
Writes a RRT manifest file ("rrt_manifest.yml") to disk
}
\description{
# FIXME: check for existence of manifest file first, and combine with old info, if any
}
\keyword{internal}

